execute bit; remove .o files
A repository should generally contain source files.  These are
defined as "the preferred form of a work for making modifications
to it.".  It shouldn't contain intermediate files generated from
source files, such as object files.  This patch removes .o files
from the repository and adds a .gitignore file to tell Git not to
try adding them again.

Python programs used for data conversion during a build process
should generally have their UNIX "executable" bit (the x in rwx)
turned on so that Make and the shell can identify them as programs.
This patch turns those on.